chore: project manager api restructure
