--- beehive/apis/project_manager.js
@@ -1,244 +1,0 @@
-import { PROJECT_API, SPRINT_API, TASK_API } from "./routes";
-
-// Create query dict to string for url parameter
-
-const getQueryString = (query) => {
-  var query_string = "";
-  for (const [key, value] of Object.entries(query)) {
-    if (!value) {
-      continue;
-    }
-    if (query_string !== "") {
-      query_string += "&";
-    }
-    query_string += `${key}=${value}`;
-  }
-  return query_string;
-};
-
-// Project APIs function
-
-export const getProjectList = async (token, query = {}) => {
-  const project_id = query["project_id"];
-  const res = await fetch(PROJECT_API + `?project_id=${project_id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const getProject = async (token, id) => {
-  const res = await fetch(PROJECT_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const createProject = async (token, body) => {
-  const res = await fetch(PROJECT_API, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const updateProject = async (token, body, id) => {
-  const res = await fetch(PROJECT_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-// Sprint APIs function
-
-export const getSprintList = async (token) => {
-  const res = await fetch(SPRINT_API, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const getSprint = async (token, id) => {
-  const res = await fetch(SPRINT_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const createSprint = async (token, body) => {
-  const res = await fetch(SPRINT_API, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const updateSprint = async (token, body, id) => {
-  const res = await fetch(SPRINT_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-// Task APIs function
-
-export const getTaskList = async (token, query) => {
-  const query_string = getQueryString(query);
-  const URL = query_string == "" ? TASK_API : `${TASK_API}?${query_string}`;
-  const res = await fetch(URL, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const getTask = async (token, id) => {
-  const res = await fetch(TASK_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const createTask = async (token, body) => {
-  const res = await fetch(TASK_API, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const updateTask = async (token, body, id) => {
-  const res = await fetch(TASK_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-// Project Task Status APIs function
-
-export const getProjectStatusList = async (token, project_id) => {
-  const res = await fetch(PROJECT_API + `/${project_id}/status`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const getProjectStatus = async (token, project_id, id) => {
-  const res = await fetch(PROJECT_STATUS_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    method: "GET",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const createProjectStatus = async (token, project_id, body) => {
-  const res = await fetch(PROJECT_STATUS_API, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
-
-export const updateProjectStatus = async (token, project_id, body, id) => {
-  const res = await fetch(PROJECT_STATUS_API + `/${id}`, {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: "Token " + token,
-    },
-    body: body,
-    method: "POST",
-  });
-  const status = await res.status;
-  const result = await res.json();
-  return [status, result];
-};
